requirements.txt
amqp==5.0.2
asgiref==3.2.10
billiard==3.6.3.0
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-haystack==3.0
django-import-export==2.4.0
django-redis==4.12.1
django-rest-multiple-models==2.1.3
django-reversion==3.0.8
django-simpleui==2021.1.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-extensions==0.6.0
drf-haystack==1.8.9
elasticsearch==2.4.1
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==5.0.2
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
prompt-toolkit==3.0.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
sentry-sdk==0.19.5
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
Whoosh==2.7.4
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
yarg==0.1.9
